refactor: 데이터 변경으로 타입 및 코드 리팩터링
