+++ fe/src/pages/GamePage/PlayerList/Player.tsx
@@ -1,23 +1,18 @@
 import { Card, CardContent } from '@/components/ui/card';
-import { PlayerProps } from '@/types/playerTypes';
 import { FaCrown } from 'react-icons/fa6';
 import VolumeBar from './VolumeBar';
+import { PlayerProps } from '@/types/roomTypes';
 
-const Player = ({
-  playerNickname,
-  isHost = false,
-  isAudioOn,
-  isReady = false,
-}: PlayerProps) => {
+const Player = ({ playerNickname, isReady }: PlayerProps) => {
   return (
     <Card className="h-full">
       <CardContent className="flex h-[4.7rem] items-center justify-between p-4">
         <div className="flex items-center gap-2">
-          {isHost && <FaCrown className="text-yellow-500" />}
+          <FaCrown className="text-yellow-500" />
           <span className="font-galmuri">{playerNickname}</span>
           {isReady && <span className="text-sm text-green-500">준비 완료</span>}
         </div>
-        <VolumeBar isOn={isAudioOn} />
+        <VolumeBar />
       </CardContent>
     </Card>
   );
