--- fe/src/types/playerTypes.ts
@@ -1,6 +1,0 @@
-export interface PlayerProps {
-  playerNickname: string;
-  isHost?: boolean;
-  isAudioOn: boolean;
-  isReady?: boolean;
-}
