+++ fe/src/types/roomTypes.ts
@@ -1,8 +1,13 @@
+export interface PlayerProps {
+  playerNickname: string;
+  isReady: boolean;
+}
+
 export interface Room {
   roomId: string;
   roomName: string;
   hostNickname: string;
-  players: string[];
+  players: PlayerProps[];
   status: 'waiting' | 'playing';
 }
